Modularize rolling checksum implementation
